Python code for Assignment 05
This is the code I submitted for my assignment 05. 
I was attempting to enhance the code with an option to remove a student from the roster, but I could never get it to work properly.
Due to the time restriction I abandoned the attemmpt.
If you like to see what I have in that enhancement I have uploaded it as well.
perhaps you can help by providing some hints on how I can acomplish it.
I am a student of (KISS: Keep It Simple Stupid), so if you get too complex my eyes will just glaze over.